update appcompat module to utilize AndroidX
