+++ gto-support-appcompat/build.gradle
@@ -2,5 +2,5 @@ dependencies {
     implementation project(':gto-support-core')
     implementation project(':gto-support-util')
 
-    implementation "com.android.support:appcompat-v7:${deps.androidSupport}"
+    implementation "androidx.appcompat:appcompat:${deps.androidXAppCompat}"
 }
