+++ gto-support-appcompat/src/main/java/androidx/appcompat/widget/AppCompatRadioButtonImpl.java
@@ -1,31 +1,31 @@
-package android.support.v7.widget;
+package androidx.appcompat.widget;
 
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.graphics.PorterDuff;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.Nullable;
-import android.support.annotation.RestrictTo;
-import android.support.v4.view.TintableBackgroundView;
 import android.util.AttributeSet;
-import android.widget.RelativeLayout;
 
-import static android.support.annotation.RestrictTo.Scope.LIBRARY_GROUP;
+import androidx.annotation.DrawableRes;
+import androidx.annotation.Nullable;
+import androidx.annotation.RestrictTo;
+import androidx.core.view.TintableBackgroundView;
+
+import static androidx.annotation.RestrictTo.Scope.LIBRARY_GROUP;
 
 @SuppressWarnings("RestrictedApi")
-public class AppCompatRelativeLayoutImpl extends RelativeLayout implements TintableBackgroundView {
+public class AppCompatRadioButtonImpl extends AppCompatRadioButton implements TintableBackgroundView {
     private AppCompatBackgroundHelper mBackgroundTintHelper;
 
-    public AppCompatRelativeLayoutImpl(final Context context) {
+    public AppCompatRadioButtonImpl(final Context context) {
         this(context, null);
     }
 
-    public AppCompatRelativeLayoutImpl(final Context context, final AttributeSet attrs) {
+    public AppCompatRadioButtonImpl(final Context context, final AttributeSet attrs) {
         this(context, attrs, 0);
     }
 
-    public AppCompatRelativeLayoutImpl(final Context context, final AttributeSet attrs, final int defStyleAttr) {
+    public AppCompatRadioButtonImpl(final Context context, final AttributeSet attrs, final int defStyleAttr) {
         super(TintContextWrapper.wrap(context), attrs, defStyleAttr);
 
         mBackgroundTintHelper = new AppCompatBackgroundHelper(this);
@@ -51,7 +51,7 @@ public class AppCompatRelativeLayoutImpl extends RelativeLayout implements Tinta
 
     /**
      * This should be accessed via
-     * {@link android.support.v4.view.ViewCompat#setBackgroundTintList(android.view.View, ColorStateList)}
+     * {@link androidx.core.view.ViewCompat#setBackgroundTintList(android.view.View, ColorStateList)}
      *
      * @hide
      */
@@ -65,7 +65,7 @@ public class AppCompatRelativeLayoutImpl extends RelativeLayout implements Tinta
 
     /**
      * This should be accessed via
-     * {@link android.support.v4.view.ViewCompat#getBackgroundTintList(android.view.View)}
+     * {@link androidx.core.view.ViewCompat#getBackgroundTintList(android.view.View)}
      *
      * @hide
      */
@@ -78,7 +78,7 @@ public class AppCompatRelativeLayoutImpl extends RelativeLayout implements Tinta
 
     /**
      * This should be accessed via
-     * {@link android.support.v4.view.ViewCompat#setBackgroundTintMode(android.view.View, PorterDuff.Mode)}
+     * {@link androidx.core.view.ViewCompat#setBackgroundTintMode(android.view.View, PorterDuff.Mode)}
      *
      * @hide
      */
@@ -92,7 +92,7 @@ public class AppCompatRelativeLayoutImpl extends RelativeLayout implements Tinta
 
     /**
      * This should be accessed via
-     * {@link android.support.v4.view.ViewCompat#getBackgroundTintMode(android.view.View)}
+     * {@link androidx.core.view.ViewCompat#getBackgroundTintMode(android.view.View)}
      *
      * @hide
      */
