+++ gto-support-appcompat/src/main/java/androidx/appcompat/widget/AppCompatRelativeLayoutImpl.java
@@ -1,30 +1,32 @@
-package android.support.v7.widget;
+package androidx.appcompat.widget;
 
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.graphics.PorterDuff;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.Nullable;
-import android.support.annotation.RestrictTo;
-import android.support.v4.view.TintableBackgroundView;
 import android.util.AttributeSet;
+import android.widget.RelativeLayout;
 
-import static android.support.annotation.RestrictTo.Scope.LIBRARY_GROUP;
+import androidx.annotation.DrawableRes;
+import androidx.annotation.Nullable;
+import androidx.annotation.RestrictTo;
+import androidx.core.view.TintableBackgroundView;
+
+import static androidx.annotation.RestrictTo.Scope.LIBRARY_GROUP;
 
 @SuppressWarnings("RestrictedApi")
-public class AppCompatRadioButtonImpl extends AppCompatRadioButton implements TintableBackgroundView {
+public class AppCompatRelativeLayoutImpl extends RelativeLayout implements TintableBackgroundView {
     private AppCompatBackgroundHelper mBackgroundTintHelper;
 
-    public AppCompatRadioButtonImpl(final Context context) {
+    public AppCompatRelativeLayoutImpl(final Context context) {
         this(context, null);
     }
 
-    public AppCompatRadioButtonImpl(final Context context, final AttributeSet attrs) {
+    public AppCompatRelativeLayoutImpl(final Context context, final AttributeSet attrs) {
         this(context, attrs, 0);
     }
 
-    public AppCompatRadioButtonImpl(final Context context, final AttributeSet attrs, final int defStyleAttr) {
+    public AppCompatRelativeLayoutImpl(final Context context, final AttributeSet attrs, final int defStyleAttr) {
         super(TintContextWrapper.wrap(context), attrs, defStyleAttr);
 
         mBackgroundTintHelper = new AppCompatBackgroundHelper(this);
@@ -50,7 +52,7 @@ public class AppCompatRadioButtonImpl extends AppCompatRadioButton implements Ti
 
     /**
      * This should be accessed via
-     * {@link android.support.v4.view.ViewCompat#setBackgroundTintList(android.view.View, ColorStateList)}
+     * {@link androidx.core.view.ViewCompat#setBackgroundTintList(android.view.View, ColorStateList)}
      *
      * @hide
      */
@@ -64,7 +66,7 @@ public class AppCompatRadioButtonImpl extends AppCompatRadioButton implements Ti
 
     /**
      * This should be accessed via
-     * {@link android.support.v4.view.ViewCompat#getBackgroundTintList(android.view.View)}
+     * {@link androidx.core.view.ViewCompat#getBackgroundTintList(android.view.View)}
      *
      * @hide
      */
@@ -77,7 +79,7 @@ public class AppCompatRadioButtonImpl extends AppCompatRadioButton implements Ti
 
     /**
      * This should be accessed via
-     * {@link android.support.v4.view.ViewCompat#setBackgroundTintMode(android.view.View, PorterDuff.Mode)}
+     * {@link androidx.core.view.ViewCompat#setBackgroundTintMode(android.view.View, PorterDuff.Mode)}
      *
      * @hide
      */
@@ -91,7 +93,7 @@ public class AppCompatRadioButtonImpl extends AppCompatRadioButton implements Ti
 
     /**
      * This should be accessed via
-     * {@link android.support.v4.view.ViewCompat#getBackgroundTintMode(android.view.View)}
+     * {@link androidx.core.view.ViewCompat#getBackgroundTintMode(android.view.View)}
      *
      * @hide
      */
