+++ gto-support-appcompat/src/main/java/androidx/appcompat/widget/ReflectionUtils.java
@@ -1,13 +1,13 @@
-package android.support.v7.widget;
-
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+package androidx.appcompat.widget;
 
 import java.lang.reflect.Field;
 
-final class ReflectionUtils {
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+public final class ReflectionUtils {
     @Nullable
-    static Field getDeclaredField(@NonNull final Class<?> type, @NonNull final String fieldName) {
+    public static Field getDeclaredField(@NonNull final Class<?> type, @NonNull final String fieldName) {
         try {
             final Field field = type.getDeclaredField(fieldName);
             field.setAccessible(true);
